consolidating abstract base class with interface
